ref(derived data): Split promotion logic to its own module
